requirements.txt
anyio==3.4.0
asgiref==3.4.1
black==21.12b0
click==8.0.3
deta==1.0.1
fastapi==0.70.1
h11==0.12.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pydantic==1.8.2
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.3
typing-extensions==4.0.1
uvicorn==0.16.0
